% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_latent.R
\name{calculate_latent}
\alias{calculate_latent}
\title{calculate latent}
\usage{
calculate_latent(data, alpha, log_p_y, log_marg_x_4d)
}
\arguments{
\item{data}{Data provided by user}

\item{alpha}{2D matrix. Numeric matrix of indicator variables that indicates adjacency of observed vs hidden variables}

\item{log_p_y}{A 2D matrix representing the log of the marginal probability of the latent variables}

\item{log_marg_x_4d}{A 4D array the contains the value of the marginal distribution for each variable, for each dimension of each hidden variable and for each data row.}
}
\value{
3D array of numerics in range (0, 1), that represent the probability for n_hidden latent y variables of dimension dim_hidden, for each observed x variable with dimensions (n_hidden, n_samples, dim_hidden)
}
\description{
Internal function to calculate the probability distribution for hidden factors for each sample
}
\details{
Calculate the probability distribution for each latent y variable given the data, the marginals and the probability of each y. This calculation is represented by Equation 16 of \strong{Greg Ver Steeg and Aram Galstyan. Maximally Informative Hierarchical Representations of High-Dimensional Data. 2015.} \url{https://arxiv.org/abs/1410.7404 }
}
