% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mis.R
\name{calculate_mis}
\alias{calculate_mis}
\title{calculate mis}
\usage{
calculate_mis(
  data,
  theta,
  marginal_description,
  log_p_y,
  p_y_given_x_3d,
  dim_visible
)
}
\arguments{
\item{data}{Data provided by user}

\item{theta}{List of estimated parameters}

\item{marginal_description}{Character string which determines the marginal distribution of the data. single marginal description applies to all variables in biocorex}

\item{log_p_y}{A 2D matrix representing the log of the marginal probability of the latent variables.}

\item{p_y_given_x_3d}{A 3D array of numerics in range (0, 1), that represent the probability that each observed x variable belongs to n_hidden latent variables of dimension dim_hidden. p_y_given_x_3d has dimensions (n_hidden, n_samples, dim_hidden)}

\item{dim_visible}{The dimension of the data provided in data - i.e. the number of discrete levels that exist in the data. Must be positive integer.}
}
\value{
Returns an array of normalised mutual information with number of columns = n_visible and number of rows = n_hidden.
}
\description{
Internal function to calculate the mutual information between each known variable and latent variable(s).
}
