% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_p_xi_given_y.R
\name{calculate_p_xi_given_y}
\alias{calculate_p_xi_given_y}
\title{calculate_p_xi_given_y}
\usage{
calculate_p_xi_given_y(x_i, thetai, marginal_description, dim_visible = NULL)
}
\arguments{
\item{x_i}{A single variable/column of data}

\item{thetai}{Estimated parameters corresponding to the single variable/column of data provided to x_i}

\item{marginal_description}{Character string which determines the marginal distribution of the data. For biocorex only 'gaussian' and 'discrete' are allowed.}

\item{dim_visible}{Integer. The dimension of the data - i.e. the number of discrete levels that exist in the data.}
}
\value{
A 3D array of marginals with dimensions: \code{(n_hidden, n_samples, dim_hidden)} - i.e. marginals for each data point in x_i given current \emph{n_hidden x dim_hidden} parameter estimates
}
\description{
Internal function to calculate marginals for a single column of data given parameter estimates
}
\details{
This function is an intermediate function with parent \code{\link{calculate_marginals_on_samples}} and child functions \code{\link{marginal_p_discrete}} and \code{\link{marginal_p_gaussian}}. It's main function is to call the appropriate child function depending on the value of the \code{marginal_description} variable.
}
