% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_p_y.R
\name{calculate_p_y}
\alias{calculate_p_y}
\title{calculate_p_y}
\usage{
calculate_p_y(p_y_given_x_3d)
}
\arguments{
\item{p_y_given_x_3d}{A 3D array of numerics in range (0, 1), that represent the probability that each observed x variable belongs to n_hidden latent variables of dimension dim_hidden. p_y_given_x_3d has dimensions (n_hidden, n_samples, dim_hidden).}
}
\value{
A 2D matrix representing the log of the marginal probability of the latent variables
}
\description{
Internal function to calculate the marginal probability of latent variables
}
\keyword{internal}
