% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_theta.R
\name{calculate_theta}
\alias{calculate_theta}
\title{calculate_theta}
\usage{
calculate_theta(
  data,
  p_y_given_x_3d,
  marginal_description,
  smooth_marginals,
  dim_visible = NULL
)
}
\arguments{
\item{data}{Data provided by user}

\item{p_y_given_x_3d}{A 3D array of numerics in range (0, 1), that represent the probability that each observed x variable belongs to n_hidden latent variables of dimension dim_hidden. p_y_given_x_3d has dimensions (n_hidden, n_samples, dim_hidden).}

\item{marginal_description}{Character string which determines the marginal distribution of the data. single marginal description applies to all variables in biocorex}

\item{smooth_marginals}{Boolean (TRUE/FALSE) which indicates whether Bayesian smoothing of marginal estimates should be used.}

\item{dim_visible}{The dimension of the data provided in data - i.e. the number of discrete levels that exist in the data. Must be positive integer.}
}
\value{
Returns a list of estimated parameters. The list has length = number of columns in the supplied \code{data}. The elements of the returned list depend on the marginal_description as follows: \itemize{ \item If the marginal description is "gaussian" a list of 2 arrays the first of which represent the estimate means, the second the estimated standard deviations. Each of these lists will have dimensions (n_hidden, dim_hidden).\item If the marginal description is "discrete", each element contains a list of length...}
}
\description{
Internal function to calculate parameters for each dimension of each hidden variable
}
\details{
This function is calls functions \code{\link{estimate_parameters_gaussian}} or \code{\link{estimate_parameters_discrete}} depending on the \code{marginal_description}.
}
\keyword{internal}
