% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_converged.R
\name{check_converged}
\alias{check_converged}
\title{check_converged}
\usage{
check_converged(tc_history, eps)
}
\arguments{
\item{tc_history}{A list recording the calculated total correlation for each loop iteration.}

\item{eps}{Defines the change in total correlation (TC) needed to signal convergence.}
}
\value{
Returns a boolean with TRUE signalling convergence and FALSE signalling no convergence.
}
\description{
Internal function to check CorEx convergence. For the first 10 iterations convergence is not checked. After this, the absolute mean difference between the 5 most recent estimations of TC, and the 5 iterations previous to that again is calculated. Convergence is considered achieved if the difference is less than the value of the eps variable.
}
