% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_maximal_run.R
\name{choose_maximal_run}
\alias{choose_maximal_run}
\title{choose_maximal_run}
\usage{
choose_maximal_run(repeat_results, repeats, max_iter)
}
\arguments{
\item{repeat_results}{List of repeated corex results}

\item{repeats}{The number of repeat runs}

\item{max_iter}{Numeric that specifies the maximum number of iterations the corex fits were run. Used in \emph{choose_maximal_run()} to detect non-converged runs.}
}
\value{
Returns the run with the highest TC of all converged runs. If no runs have converged, \emph{choose_maximal_run()} returns the run with maximal TC anyway - in such cases the results should not be interpreted, but may be useful for diagnostic purposes.
}
\description{
Internal function to select the optimal corex run when biocroex is run with repeats > 1
}
\keyword{internal}
