% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corex_BIC.R
\name{corex_BIC}
\alias{corex_BIC}
\title{corex_BIC}
\usage{
corex_BIC(corex_obj)
}
\arguments{
\item{corex_obj}{A fitted corex object}
}
\value{
numeric that contains BIC value
}
\description{
Experimental BIC calculation
}
