% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corex_RMSE.R
\name{corex_RMSE}
\alias{corex_RMSE}
\title{corex_RMSE}
\usage{
corex_RMSE(corex_obj, new_data = NULL)
}
\arguments{
\item{corex_obj}{A fitted corex object}

\item{new_data}{Optionally provided new data. Must contain same variables as fit data. If new_data not provided uses fit data}
}
\value{
numeric that represents the RMSE
}
\description{
Experimental root-mean-squared-error calculation
}
