% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicorex.R
\name{epicorex}
\alias{epicorex}
\title{epicorex}
\usage{
epicorex(
  data,
  n_hidden = 1,
  dim_hidden = 2,
  marginal_description,
  smooth_marginals = FALSE,
  minmarg = NULL,
  eps = 1e-06,
  verbose = FALSE,
  repeats = 1,
  return_all_runs = FALSE,
  max_iter = 100
)
}
\arguments{
\item{data}{Data provided by user. Allows for mixed data types. Eg categorical or binomial or continuous or discrete in same dataset}

\item{n_hidden}{An integer number of hidden variables to search for. Default = 1.}

\item{dim_hidden}{Each hidden unit can take \code{dim_hidden} discrete values. Default = 2}

\item{marginal_description}{Character string which determines the marginal distribution of the data.For epicorex, marginal_description must be a vector of strings of length equal to the number of columns in \code{data}. Allowable marginal descriptions are: gaussian, discrete, bernoulii currently - more may be added later.}

\item{smooth_marginals}{Boolean (TRUE/FALSE) which indicates whether Bayesian smoothing of marginal estimates should be used.
#' @param minmarg EXPERIMENTAL If NULL has no effect. If not NULL should be a negative number that places a minimal value on individual log_marginal values. This is an alternative approach to imposing a minimum value on estimated standard deviations when using the gaussian marginal distribution, with the aim to prevent errors cause by dividing by zero which can occur in some circumstances}

\item{eps}{The maximal change in TC across 10 iterations needed signal convergence}

\item{verbose}{Default FALSE. If TRUE, epicorex feeds back to user the iteration count and TCS each iteration. Useful to see progression if fitting a larger dataset.}

\item{repeats}{How many times to run epicorex on the data using random initial values. Corex will return the run which leads to the maximum TC. Default is 1. For a new dataset, recommend to leave it as 1 to see how long epicorex takes, however for more trustworthy results a higher numbers recommended (e.g. 25).}

\item{return_all_runs}{Default FALSE. If FALSE epicorex returns a single object of class rcorex. If TRUE epicorex returns all runs of epicorex as a list - the length of which = \code{repeats}. In this case the returned results are not rcorex objects, but have the same components of an rcorex object with class list.}

\item{max_iter}{numeric. Maximum number of iterations before ending. Default = 100}
}
\value{
Returns either a rcorex object or a list of repeated runs as determined by the \code{return_all_runs} argument. An rcorex object is a list that contains the following components:
#' \enumerate{
\item{data - the user data supplied in call to corex.}
\item{call - the call used to run corex.}
\item{tcs - a vector of TC for n_hidden variables.}
\item{alpha - a 2D adjaceny matrix of connections between input variables and hidden variables.}
\item{p_y_given_x - a 3D array of numerics in range (0, 1), that represent the probability that each observed x variable belongs to n_hidden latent variables of dimension dim_hidden. p_y_given_x has dimensions (n_hidden, n_samples, dim_hidden).}
\item{theta - a list of the estimated parameters}
\item{log_p_y - a 2D matrix representing the log of the marginal probability of the latent variables.}
\item{log_z - a 2D matrix containing the pointwise estimate of total correlation explained by each latent variable for each sample - this is used to estimate overall total correlation.}
\item{dim_visible - only present if discrete marginals were specified. Lists the number of discrete levels that exist in the data.}
\item{iterations - the number of iterations for which the algorithm ran.}
\item{tc_history - a list that records the TC results for each iteration of the algorithm.}
\item{marginal_description - a character string which determines the marginal distribution of the data. }
\item{mis - an array that specifies the mutual information between each observed variable and hidden variable.}
\item{clusters - a vector that assigns a hidden variable label to each input variable.}
\item{labels - a 2D matrix of dimensions \code{(nrow(data), n_hidden)} that assigns a dimension label for each hidden variable to each row of data.}
}
}
\description{
Port of Python epicorex function
}
\details{
This function is an extension of the biocorex function to allow for mixed data-types as typically found in epidemiological datasets - e.g. categorical data + continuous data for example.
}
