% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters_discrete.R
\name{estimate_parameters_discrete}
\alias{estimate_parameters_discrete}
\title{estimate_parameters_discrete}
\usage{
estimate_parameters_discrete(
  x_i,
  p_y_given_x_3d,
  smooth_marginals,
  dim_visible
)
}
\arguments{
\item{x_i}{A single variable/column of data}

\item{p_y_given_x_3d}{A 3D array of numerics in range (0, 1), that represent the probability that each observed x variable belongs to n_hidden latent variables of dimension dim_hidden. p_y_given_x_3d has dimensions (n_hidden, n_samples, dim_hidden).}

\item{smooth_marginals}{Boolean (TRUE/FALSE) which indicates whether Bayesian smoothing of marginal estimates should be used as described in \strong{Pepke, S., Ver Steeg, G. Comprehensive discovery of subsample gene expression components by information explanation: therapeutic implications in cancer. BMC Med Genomics 10, 12 (2017).} \url{https://doi.org/10.1186/s12920-017-0245-6}}

\item{dim_visible}{The dimension of the data provided in data - i.e. the number of discrete levels that exist in the data. Must be positive integer.}
}
\value{
Returns a 3D array of dimensions (dim_visible, n_hidden, dim_hidden) that represent the probability of each dim_visible occurring for each dim_hidden across each n_hidden.
}
\description{
Internal function to estimate parameters for data with discrete marginal description
}
\details{
Estimates the probability of each dim_visible occurring for each dim_hidden across each n_hidden.
}
\keyword{internal}
