% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters_gaussian.R
\name{estimate_parameters_gaussian}
\alias{estimate_parameters_gaussian}
\title{estimate_parameters_gaussian}
\usage{
estimate_parameters_gaussian(x_i, p_y_given_x_3d, smooth_marginals)
}
\arguments{
\item{x_i}{A single variable/column of data}

\item{p_y_given_x_3d}{A 3D array of numerics in range (0, 1), that represent the probability that each observed x variable belongs to n_hidden latent variables of dimension dim_hidden. p_y_given_x_3d has dimensions (n_hidden, n_samples, dim_hidden).}

\item{smooth_marginals}{Boolean (TRUE/FALSE) which indicates whether Bayesian smoothing of marginal estimates should be used as described in \strong{Pepke, S., Ver Steeg, G. Comprehensive discovery of subsample gene expression components by information explanation: therapeutic implications in cancer. BMC Med Genomics 10, 12 (2017).} \url{https://doi.org/10.1186/s12920-017-0245-6}}
}
\value{
Returns a list of 2 matrices each of dimension (n_hidden, dim_hidden). The first matrix represents estimated means for n_hidden x dim_hidden parameters, while the second matrix represents corresponding standard deviations.
}
\description{
Internal function to estimate parameters for data with gaussian marginal description
}
\keyword{internal}
