% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_se.R
\name{estimate_se}
\alias{estimate_se}
\title{estimate_se}
\usage{
estimate_se(x_i, p_y_given_x_3d, num_obs, reps = 20)
}
\arguments{
\item{x_i}{A single variable/column of data}

\item{p_y_given_x_3d}{A 3D array of numerics in range (0, 1), that represent the probability that each observed x variable belongs to n_hidden latent variables of dimension dim_hidden. p_y_given_x_3d has dimensions (n_hidden, n_samples, dim_hidden).}

\item{num_obs}{A 2D array of dimensions (n_hidden, dim_hidden) representing expected counts of n_hidden x dim_hidden summed over data samples (rows) - can be non-integer}

\item{reps}{numeric to specify number of bootstrap estimates to calculate. Default = 20}
}
\value{
Returns a list of 4 numerics names m1, m2, se1, se2 which are used in smoothing calculations by \code{\link{estimate_parameters_gaussian}}
}
\description{
Internal Function to calculate bootstrap estimate of mean and standard error of gaussian parameters
}
