% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sig.R
\name{estimate_sig}
\alias{estimate_sig}
\title{estimate_sig}
\usage{
estimate_sig(x_select, p_y_given_x_3d, term)
}
\arguments{
\item{x_select}{Selected data}

\item{p_y_given_x_3d}{A 3D array of numerics in range (0, 1), that represent the probability that each observed x variable belongs to n_hidden latent variables of dimension dim_hidden. p_y_given_x_3d has dimensions (n_hidden, n_samples, dim_hidden).}

\item{term}{Intermediate term calculated by parent function}
}
\value{
REturns bootstrapped estimates
}
\description{
Internal Function to calculate bootstrap estimate of mean and standard error of gaussian parameters
}
