% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dimvisible.R
\name{get_dimvisible}
\alias{get_dimvisible}
\title{Get dim_visible from supplied data}
\usage{
get_dimvisible(data, max_dim = 31)
}
\arguments{
\item{data}{Data provided by user. For biocorex data can either be continuous (gaussian) or discrete (consectutive integers 0, 1, 2, 3...etc). Data types cannot by mixed in this implementation.}

\item{max_dim}{Sets the maximum dimension allowed for discrete data. Default = 31}
}
\value{
An integer that represents the dimension of unique values in the discrete data.
}
\description{
Internal function to determine the dimension of supplied discrete data
}
\details{
This function determines the dimension of unique values in the supplied data.
}
\keyword{internal}
