% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dimvisible_byvar.R
\name{get_dimvisible_byvar}
\alias{get_dimvisible_byvar}
\title{get_dimvisible_byvar from supplied data}
\usage{
get_dimvisible_byvar(data, marginal_description)
}
\arguments{
\item{data}{Data provided by user. For biocorex data can either be continuous (gaussian) or discrete (consectutive integers 0, 1, 2, 3...etc). Data types cannot by mixed in this implementation.}

\item{marginal_description}{Character string which determines the marginal distribution of the data. Should be of length equal to the number of variables in data. get_dimvisible_byvar() calculates the dimension when the marginal_description = 'discrete'}
}
\value{
An integer that represents the dimension of unique values in the discrete data.
}
\description{
Internal function to determine the dimension of supplied discrete by column
}
\details{
This function determines the dimension of unique values in the supplied data for each column in the supplied data frame, if the marginal description = discrete for a given variable.
}
\keyword{internal}
