% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearcorex.R
\name{linearcorex}
\alias{linearcorex}
\title{Fit linearcorex to a dataset}
\usage{
linearcorex(
  data,
  n_hidden = 10,
  max_iter = 10000,
  tol = 1e-05,
  anneal = TRUE,
  discourage_overlap = TRUE,
  gaussianize = "standard",
  verbose = FALSE
)
}
\arguments{
\item{data}{Data provided by user.}

\item{n_hidden}{An integer number of hidden variables to search for. Default = 10.}

\item{max_iter}{numeric. Maximum number of iterations. Default = 10000}

\item{tol}{Tolerance}

\item{anneal}{Default = TRUE}

\item{discourage_overlap}{Default = TRUE}

\item{gaussianize}{Default = standard}

\item{verbose}{Default = FALSE}
}
\value{
Returns linearcorex fit objects as a list
}
\description{
EXPERIMENTAL implmentation of linearcorex algorithm
}
\details{
Experimental implementation of  \url{https://github.com/gregversteeg/LinearCorex}. Currently not fully tested.
}
