% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_p_bernoulli.R
\name{marginal_p_bernoulli}
\alias{marginal_p_bernoulli}
\title{Marginal Probabilities for binary data}
\usage{
marginal_p_bernoulli(x_i, thetai)
}
\arguments{
\item{x_i}{A single variable/column of data}

\item{thetai}{Estimated parameters corresponding to the single variable/column of data provided to x_i}
}
\value{
A three dimensional array of marginals with dimensions: \code{(n_hidden, n_samples, dim_hidden)} - i.e. marginals for each data point in x_i given current \emph{n_hidden x dim_hidden} parameter estimates
}
\description{
Internal function to estimate the marginals for binary (bernoulli) data.
}
\details{
Estimate the marginals for a single column of data \code{x_i} for each hidden variable with dimension dim_hidden. For discrete data this can be calculated directly from the data.
}
