% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_p_discrete.R
\name{marginal_p_discrete}
\alias{marginal_p_discrete}
\title{Marginal Probabilities for discrete data}
\usage{
marginal_p_discrete(x_i, thetai, dim_visible)
}
\arguments{
\item{x_i}{A single variable/column of data}

\item{thetai}{Estimated parameters corresponding to the single variable/column of data provided to x_i}

\item{dim_visible}{The dimension of the variable provided in x_i - i.e. the number of discrete levels}
}
\value{
A three dimensional array of marginals with dimensions: \code{(n_hidden, n_samples, dim_hidden)} - i.e. marginals for each data point in x_i given current \emph{n_hidden x dim_hidden} parameter estimates
}
\description{
Internal function to estimate the marginals for discrete data.
}
\details{
Estimate the marginals for a single column of data \code{x_i} for each hidden variable with dimension dim_hidden. For discrete data this can be calculated directly from the data.
}
\keyword{internal}
