% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_p_gaussian.R
\name{marginal_p_gaussian}
\alias{marginal_p_gaussian}
\title{Marginal Probabilities for Gaussian distributed data}
\usage{
marginal_p_gaussian(x_i, thetai)
}
\arguments{
\item{x_i}{A single variable/column of data}

\item{thetai}{Estimated parameters corresponding to the single variable/column of data provided to x_i}
}
\value{
A three dimensional array of marginals with dimensions: \code{(n_hidden, dim_hidden, n_samples)} - i.e. marginals for each data point in x_i given current \emph{n_hidden x dim_hidden} parameter estimates
}
\description{
Internal function to estimate the marginals for Gaussian distributed data.
}
\details{
Estimate the marginals for a single column of data \code{x_i} for each hidden variable with dimension dim_hidden. For continuous data we estimate \eqn{p\left(y_{j} \mid x_{i}\right) / p\left(y_{j}\right)} indirectly by using Bayes rule to rewrite this as \eqn{p\left(x_{i} \mid y_{j}\right) / p\left(x_{i}\right)}, which can be estimated.
}
\keyword{internal}
