% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_bootstrap.R
\name{mi_bootstrap}
\alias{mi_bootstrap}
\title{mi_bootstrap}
\usage{
mi_bootstrap(
  data,
  marginal_description,
  minmarg,
  theta,
  log_p_y,
  p_y_given_x_3d,
  dim_visible,
  smooth_marginals,
  n_permutes = 20
)
}
\arguments{
\item{data}{Data provided by user}

\item{marginal_description}{Character string which determines the marginal distribution of the data.}

\item{theta}{List of estimated parameters}

\item{log_p_y}{A 2D matrix representing the log of the marginal probability of the latent variables}

\item{p_y_given_x_3d}{A 3D array of numerics in range (0, 1), that represent the probability that each observed x variable belongs to n_hidden latent variables of dimension dim_hidden. p_y_given_x_3d has dimensions (n_hidden, n_samples, dim_hidden).}

\item{dim_visible}{The dimension of the data provided in data - i.e. the number of discrete levels that exist in the data. Must be positive integer.}

\item{smooth_marginals}{Boolean (TRUE/FALSE) which indicates whether Bayesian smoothing of marginal estimates should be used.}

\item{n_permutes}{numeric to specify number of bootstrap estimates to calculate. Default = 20}
}
\value{
Returns a list
}
\description{
Internal function to calculate bias and sd of MIS via bootstrap
}
\keyword{internal}
