% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_latent.R
\name{normalise_latent}
\alias{normalise_latent}
\title{normalise_latent}
\usage{
normalise_latent(log_p_y_given_x_unnorm_3d)
}
\arguments{
\item{log_p_y_given_x_unnorm_3d}{A 3D array with dimensions (n_hidden, n_samples, dim_hidden) containing the un-normalised latent variable distribution.}
}
\value{
Returns a list called out which contains two elements:
\enumerate{
\item {A 3D array of numerics in range (0, 1), that represent the probability that each observed x variable belongs to n_hidden latent variables of dimension dim_hidden. p_y_given_x_3d has dimensions (n_hidden, n_samples, dim_hidden).}
\item{A 2D matrix containing the pointwise estimate of total correlation explained by each latent variable for each sample.}
}
}
\description{
Internal function to normalise the latent variable distribution
}
