% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.rcorex.R
\name{print.summary.rcorex}
\alias{print.summary.rcorex}
\title{Print a summary.rcorex object}
\usage{
\method{print}{summary.rcorex}(x, ...)
}
\arguments{
\item{x}{A object of class summary.rcorex}

\item{...}{Not used}
}
\description{
Print method for a summary.rcorex object
}
