% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_theta_epi.R
\name{reorder_theta_epi}
\alias{reorder_theta_epi}
\title{reorder_theta_epi}
\usage{
reorder_theta_epi(theta, marginal_description, ord)
}
\arguments{
\item{theta}{List of estimated parameters.}

\item{marginal_description}{Character string which determines the marginal distribution of the data. single marginal description applies to all variables in biocorex}

\item{ord}{Vector of integers that give desired reorder sequence}
}
\value{
Returns a new version of theta reordered according to the pattern supplied to \code{ord}
}
\description{
Internal helper function to reorder the theta parameters list for the epicorex function.
}
\details{
This function is necessary as the structure of theta varies for each parameter type. THis function is used after convergence to help package results for return to user. Is it called by \code{\link{sort_results}} which passes the desired order via the \code{ord} parameter.
}
