% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_results.R
\name{sort_results}
\alias{sort_results}
\title{sort_results}
\usage{
sort_results(
  data,
  cl,
  n_hidden,
  dim_visible,
  marginal_description,
  smooth_marginals,
  tcs,
  alpha,
  p_y_given_x_3d,
  theta,
  log_p_y,
  log_z,
  tc_history,
  names,
  state
)
}
\arguments{
\item{data}{Data provided by user}

\item{cl}{User call to biocorex}

\item{n_hidden}{An integer number of hidden variables to search for.}

\item{dim_visible}{The dimension of the data provided when discrete marginal distribution is specified - i.e. the number of discrete levels that exist in the data. Must be positive integer.}

\item{marginal_description}{Character string which determines the marginal distribution of the data. single marginal description applies to all variables in biocorex.}

\item{smooth_marginals}{Boolean (TRUE/FALSE) which indicates whether Bayesian smoothing of marginal estimates should be used.}

\item{tcs}{Vector of length n_hidden - contains the TC for each hidden factor. This is used to decided the sort order for all the other parameters such that hidden factors are returned to used in order of largest TC to smallest TC.}

\item{alpha}{Adjacency matrix between input variables and hidden variables. In range [0,1].}

\item{p_y_given_x_3d}{A 3D array of numerics in range (0, 1), that represent the probability of n_hidden latent variables of dimension dim_hidden, for each observed x variable with dimensions (n_hidden, n_samples, dim_hidden)}

\item{theta}{List of estimated parameters}

\item{log_p_y}{A 2D matrix representing the log of the marginal probability of the latent variables.}

\item{log_z}{A 2D matrix containing the pointwise estimate of total correlation explained by each latent variable for each sample - this is used to estimate overall total correlation.}

\item{tc_history}{A list that records the TC results for each iteration of the algorithm. Used to calculate if convergence has been reached.}

\item{names}{A vector of the variables names of the supplied data.}

\item{state}{A string that describes the final state of corex (i.e. "Converged", "Negative tcs", "Unconverged").}
}
\value{
Returns list of corex algorthim results sorted in descending order by TC of the latent variables. The list includes the following elements:
\enumerate{
\item{data - the user data supplied in call to corex.}
\item{call - the call used to run corex.}
\item{tcs - a vector of TC for n_hidden variables.}
\item{alpha - a 2D adjaceny matrix of connections between input variables and hidden variables.}
\item{p_y_given_x - a 3D array of numerics in range (0, 1), that represent the probability that each observed x variable belongs to n_hidden latent variables of dimension dim_hidden. p_y_given_x has dimensions (n_hidden, n_samples, dim_hidden).}
\item{theta - a list of the estimated parameters}
\item{log_p_y - a 2D matrix representing the log of the marginal probability of the latent variables.}
\item{log_z - a 2D matrix containing the pointwise estimate of total correlation explained by each latent variable for each sample - this is used to estimate overall total correlation.}
\item{dim_visible - only present if discrete marginals were specified. Lists the number of discrete levels that exist in the data.}
\item{iterations - the number of iterations for which the algorithm ran.}
\item{tc_history - a list that records the TC results for each iteration of the algorithm.}
\item{marginal_description - a character string which determines the marginal distribution of the data. }
\item{mis - an array that specifies the mutual information between each observed variable and hidden variable.}
\item{clusters - a vector that assigns a hidden variable label to each input variable.}
\item{labels - a 2D matrix of dimensions \code{(nrow(data), n_hidden)} that assigns a dimension label for each hidden variable to each row of data.}
}
}
\description{
Internal function to sort biocorex results for output to user.
}
\keyword{internal}
