% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rcorex.R
\name{summary.rcorex}
\alias{summary.rcorex}
\title{Summarise an rcorex object}
\usage{
\method{summary}{rcorex}(object, ...)
}
\arguments{
\item{object}{An object of class rcorex}

\item{...}{Not used}
}
\value{
Returns a list with the following model summary parameters:
\enumerate{
\item{call - the call used to run corex.}
\item{datadim - a vector enumerating the number of rows and columns in the data.}
\item{state - whether the model has converged or not.}
\item{iters - the number of iterations carried out by rcorex}
\item{tcs - a vector of TC for n_hidden variables.}
}
}
\description{
Summary method for an rcorex object
}
