% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_corex.R
\name{transform_corex}
\alias{transform_corex}
\title{transform_corex}
\usage{
transform_corex(corex_obj, new_data, details = FALSE)
}
\arguments{
\item{corex_obj}{A fit biocorex object of class 'rcorex'.}

\item{new_data}{New data to transform into cluster labels.}

\item{details}{Boolean. Default is FALSE - returns cluster labels. If TRUE returns list of 2 arrays:
\enumerate{
\item{p_y_given_x_3d}
\item{log_z}
}}
}
\description{
Function to transform new_data to corex labels using fit corex object
}
\details{
Experimental!
}
