% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_info.R
\name{unique_info}
\alias{unique_info}
\title{unique_info}
\usage{
unique_info(correct)
}
\arguments{
\item{correct}{Index of correct predictions of hidden variable assignment obtained by cross-referencing max_indexes of p_y_given_x_3d and max_indices of log_marg_x_4d for each possible dim_hidden. It indicates whether the estimate based on x_i for y_j is correct for sample l.}
}
\value{
Returns a vector of length n_hidden the represents an estimate of fraction of unique info in each n_hidden
}
\description{
Internal function
}
\keyword{internal}
