% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tc.R
\name{update_tc}
\alias{update_tc}
\title{Internal function to calculate marginal probility of latent variables}
\usage{
update_tc(log_z, tcs, tc_history)
}
\arguments{
\item{log_z}{A 2D matrix containing the pointwise estimate of total correlation explained by each latent variable for each sample.}

\item{tcs}{The total correlation explained by each latent variable in unit of nats (natural unit of information)}

\item{tc_history}{A log of the estimated total correlation with each iteration of the algorithm.}
}
\value{
log_p_y. The log of the marginal probability of the latent variables.
}
\description{
Internal function to calculate marginal probility of latent variables
}
\keyword{internal}
