% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_marginals_on_samples.R
\name{calculate_marginals_on_samples}
\alias{calculate_marginals_on_samples}
\title{Calculate marginals on data}
\usage{
calculate_marginals_on_samples(
  data,
  theta,
  marginal_description,
  log_p_y,
  dim_visible = NULL
)
}
\arguments{
\item{data}{Data provided by user}

\item{theta}{List of estimated parameters}

\item{marginal_description}{Character string which determines the marginal distribution of the data. A single marginal description applies to all variables in biocorex.}

\item{log_p_y}{A 2D matrix representing the log of the marginal probability of the latent variables}

\item{dim_visible}{The dimension of the data provided - i.e. the number of discrete levels that exist in the data. Must be positive integer.}
}
\value{
4D array of dimensions: \code{(n_hidden, n_samples, n_visible, dim_hidden )} where n_samples is the number of rows in the user provided data, and n_visible is the number of columns. Returned data is result fo the calculation \eqn{log \left( \frac{p\left(y_{j} \mid x_{i}\right)}{p\left(y_{j}\right)} \right))} for each j,sample,i,y_j
}
\description{
Internal function to calculate marginals given data and parameter estimates
}
\details{
Calculate the value of the marginal distribution for each variable, for each dimension of each hidden variable and each sample
}
\keyword{internal}
