% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rcorex.R
\name{plot.rcorex}
\alias{plot.rcorex}
\title{Plot TC vs number of iterations for an rcorex object}
\usage{
\method{plot}{rcorex}(x, ...)
}
\arguments{
\item{x}{A fit corex model object}

\item{...}{Not used}
}
\value{
A ggplot2 graph object
}
\description{
Function to plot corex TC per hidden factor vs iterations. Visualises algorithm convergence.
}
